feat: add Solana Summit Germany pages
