+++ src/app/sitemap.ts
@@ -37,13 +37,13 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       priority: 0.8,
     },
     {
-      url: `${baseUrl}/solana-summit`,
+      url: `${baseUrl}/solana-summit-germany`,
       lastModified: new Date(),
       changeFrequency: "weekly" as const,
       priority: 0.9,
     },
     {
-      url: `${baseUrl}/solana-summit/agenda`,
+      url: `${baseUrl}/solana-summit-germany/agenda`,
       lastModified: new Date(),
       changeFrequency: "weekly" as const,
       priority: 0.8,
