+++ src/sections/solana-summit/data.ts
@@ -88,7 +88,7 @@ export const agendaItems: AgendaItem[] = [
     time: "11:00 - 11:10",
     title: "Opening Keynote",
     type: "Keynote",
-    speakers: ["Patricia Albrecht, Superteam Germany"],
+    speakers: ["Patricia Albrecht (Superteam Germany)"],
     location: "Main Stage",
     section: "Opening",
     status: "confirmed",
@@ -97,7 +97,7 @@ export const agendaItems: AgendaItem[] = [
     time: "11:10 - 11:20",
     title: "The State of Solana",
     type: "Lightning Talk",
-    speakers: ["Mert Mumtaz, CEO & Co-Founder, Helius"],
+    speakers: ["Mert Mumtaz (Helius)"],
     location: "Main Stage",
     section: "Opening",
     status: "confirmed",
@@ -106,7 +106,10 @@ export const agendaItems: AgendaItem[] = [
     time: "11:20 - 11:35",
     title: "Jupiter: Building the Everything App",
     type: "Fireside Chat",
-    speakers: ["Xiao-Xiao, President, Jupiter", "Patricia Albrecht"],
+    speakers: [
+      "Xiao-Xiao Zhu (Jupiter)",
+      "Patricia Albrecht (Superteam Germany)",
+    ],
     location: "Main Stage",
     section: "Opening",
     status: "confirmed",
@@ -187,7 +190,7 @@ export const agendaItems: AgendaItem[] = [
     time: "18:25 - 18:30",
     title: "Closing Statement",
     type: "Keynote",
-    speakers: ["Aditya"],
+    speakers: ["Aditya Shetty (Superteam)"],
     location: "Main Stage",
     section: "Day Close",
   },
